Enable SRC functions on progress code events (#224)
As functions 11, 12, 13 are to display SRCs & hexwords of any type
(say progress SRC/ PEL SRC), whichever SRC event comes first these
functions has to be enabled.

In the existing design the functions 11, 12, 13 are enabled only when
PEL SRC events are received.

This commit enables the SRC functions 11, 12 and 13 even when progress
code events are received by ibm-panel application.

Test:
Tested on P10 rainier to see even if PEL is not found, on occurrence of
progress code events, function 11, 12, 13 are enabled.

Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>